Impl navigate to BottomSheet screen
