+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListFragment.kt
@@ -19,7 +19,6 @@ import com.numero.material_gallery.components.DesignComponent
 import com.numero.material_gallery.components.appbar.bottom.BottomAppBarTypeActivity
 import com.numero.material_gallery.components.appbar.top.TopAppBarTypeActivity
 import com.numero.material_gallery.components.bottomnavigation.BottomNavigationActivity
-import com.numero.material_gallery.components.sheet.BottomSheetActivity
 import com.numero.material_gallery.view.ListItemAdapter
 import kotlinx.android.synthetic.main.fragment_component_list.*
 
@@ -113,7 +112,7 @@ class ComponentListFragment : Fragment(R.layout.fragment_component_list) {
                 findNavController().navigate(R.id.action_ComponentList_to_ModalBottomSheet)
             }
             DesignComponent.BOTTOM_SHEET -> {
-                startActivity(BottomSheetActivity.createIntent(requireContext()))
+                findNavController().navigate(R.id.action_ComponentList_to_BottomSheet)
             }
             DesignComponent.CHECKBOX -> {
                 findNavController().navigate(R.id.action_ComponentList_to_Checkbox)
