--- app/src/main/java/com/numero/material_gallery/components/sheet/BottomSheetActivity.kt
@@ -1,75 +1,0 @@
-package com.numero.material_gallery.components.sheet
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import android.view.View
-import android.widget.LinearLayout
-import androidx.appcompat.app.AppCompatActivity
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.numero.material_gallery.R
-import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.ConfigRepository
-import kotlinx.android.synthetic.main.activity_bottom_sheet.*
-import org.koin.android.ext.android.inject
-
-class BottomSheetActivity : AppCompatActivity(R.layout.activity_bottom_sheet) {
-
-    private val configRepository by inject<ConfigRepository>()
-
-    private lateinit var behavior: BottomSheetBehavior<LinearLayout>
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        setTheme(configRepository.themeRes)
-        super.onCreate(savedInstanceState)
-        setSupportActionBar(toolbar)
-
-        supportActionBar?.apply {
-            setDisplayHomeAsUpEnabled(true)
-        }
-
-        showBottomSheetButton.setOnClickListener {
-            behavior.state = BottomSheetBehavior.STATE_EXPANDED
-        }
-        draggableSwitch.setOnCheckedChangeListener { _, isChecked ->
-            behavior.isDraggable = isChecked
-        }
-
-        behavior = BottomSheetBehavior.from(bottomSheetLayout).apply {
-            state = BottomSheetBehavior.STATE_HIDDEN
-            addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
-                override fun onSlide(p0: View, p1: Float) {
-                }
-
-                override fun onStateChanged(view: View, state: Int) {
-                    showBottomSheetButton.isEnabled = state == BottomSheetBehavior.STATE_HIDDEN
-                }
-            })
-        }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        menuInflater.inflate(R.menu.menu_common, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
-            R.id.action_info -> {
-                ThemeInfoBottomSheetDialog.newInstance().showIfNeeded(supportFragmentManager)
-                true
-            }
-            android.R.id.home -> {
-                finish()
-                true
-            }
-            else -> super.onOptionsItemSelected(item)
-        }
-    }
-
-    companion object {
-        fun createIntent(context: Context): Intent = Intent(context, BottomSheetActivity::class.java)
-    }
-}
