+++ app/src/main/res/navigation/nav_graph.xml
@@ -14,6 +14,9 @@
             android:id="@+id/action_ComponentList_to_Settings"
             app:destination="@id/SettingsScreen" />
         <action
+            android:id="@+id/action_ComponentList_to_BottomSheet"
+            app:destination="@id/BottomSheetScreen" />
+        <action
             android:id="@+id/action_ComponentList_to_Chip"
             app:destination="@id/ChipScreen" />
         <action
@@ -71,6 +74,11 @@
             android:id="@+id/action_ComponentList_to_TextField"
             app:destination="@id/TextFieldScreen" />
     </fragment>
+
+    <fragment
+        android:id="@+id/BottomSheetScreen"
+        android:name="com.numero.material_gallery.components.sheet.BottomSheetFragment"
+        tools:layout="@layout/fragment_bottom_sheet" />
 
     <fragment
         android:id="@+id/ChipScreen"
